pyproject.toml
[tool.poetry]
name = "omoide"
version = "0.1.0"
description = "Media content storing and filtering system"
authors = ["IgorZyktin <nicord@yandex.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.18"
Pillow = "^8.2.0"
colorama = "^0.4.4"
Flask = "^2.0.1"
pip = "^21.1.3"
pydantic = "^1.8.2"
click = "^8.0.1"
ujson = "^4.1.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
